changes based on the feedback received in the pull request review
