+++ Frontend/src/views/MainLayout.vue
@@ -1,56 +1,15 @@
 <script setup lang="ts">
-
-import axios from "axios";
 import CommunitySidebar from "@/components/CommunitySidebar.vue";
-
-async function getRequest() {
-
-    console.log('get request greetings')
-
-    const res = await axios.get('https://localhost/api/comms', {
-        params: {
-            page: 1
-        },
-        headers: {
-            'accept': 'application/ld+json'
-        }
-    })
-        .then(response => {
-            // Handle the response
-            console.log(response.data);
-        })
-        .catch(error => {
-            // Handle errors
-            console.error('Error:', error);
-        });
-}
-
-async function postRequest() {
-    console.log('post request')
-
-    const form = new FormData()
-    form.append('name', "nowe imie")
-    const res = await axios.post('https://localhost/greetings', form, {
-        headers: {
-            "Content-Type": "application/ld+json"
-        }
-    })
-    console.log(res)
-    console.log(res.data)
-}
-
 </script>
 
 <template>
-    <main class="text-white mt-12 flex justify-center">
+    <main class="text-white flex justify-center">
         <div class="flex min-h-screen w-4/5 gap-32">
             <div class="z-20 bg-main-bg hidden xl:block">
                 <CommunitySidebar/>
             </div>
 
             <div class="grow flex flex-col gap-4 items-center scroll-smooth">
-<!--                <button class="bg-blue-500 rounded-md p-2" @click="getRequest">Communities</button>-->
-<!--                <button class="bg-blue-500 rounded-md p-2" @click="postRequest">not fixed btw</button>-->
                 <router-link :to="{name: 'community', params: { community: 'livestreamfail'}}">
                     <button class="bg-blue-500 rounded-md p-2">Community</button>
                 </router-link>
